Use double precision time in scene

+++ src/scene/Scene.hpp
@@ -3,5 +3,5 @@
 class Scene
 {
 public:
-	virtual void Update(float deltaTime) = 0;
+	virtual void Update(double deltaTime) = 0;
 };
